requirements.txt
streamlit
pillow
opencv-python
roboflow
inference-sdk
